refactor: extract applyNodeLinkFilter utility from equalizer and compressor routes
Both routes contained a near-identical block of code that checks for
a live NodeLink player connection and applies audio filters via
player.node.rest.updatePlayer(). Extract this into a single shared
function in lib/applyNodeLinkFilter.ts, reducing ~46 duplicated lines
across two files while preserving identical behavior.
